ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]

sqlmodel
psycopg2-binary
python-dotenv
apscheduler
httpx
pandas==2.2.2


langchain-core
langchain-community
langchain-experimental
langchain                
rapidfuzz                
promptfoo 

langchain-ollama
